Combined Codes, Now included two methods of Analysis
Codes can be tested in the test.ipynb file, can be read by vscode or jupyter notebook. The sort_csv.py is to sort scale the data set. The process_functions.py is to process and visualize results from two analysis. One analysis is from the previous group. I call it Spectral analysis. The second method use least square method to predict isotopes composition in the unknown. So I call it Composition analysis