upload modified code
complete error propagation calculation for multi degression. Works remain: data collection for spectral peaks of each isotopes. need a function to compare these peaks and output the isotopes with the most similar peak energy comparing with the unknown, together with uncertainty of each comparison. That will be it for the spectral analysis.